+++ src/view/login/LoginMenu/LoginMenu.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Button from "@/common/component/Button";
-import { LoginButtonStyle } from "@/shared/component/Header/LoginMenu.css";
 import { buttonContainer } from "@/shared/component/Header/index.css";
+import { LoginButtonStyle } from "@/view/login/LoginMenu/LoginMenu.css";
 import { useRouter } from "next/navigation";
 
 const LoginMenu = () => {
@@ -12,7 +12,6 @@ const LoginMenu = () => {
 
   return (
     <div className={`${buttonContainer} ${LoginButtonStyle}`}>
-      {/** 임시 버튼 */}
       <Button aria-label="Login" size="small" onClick={handleProfileClick}>
         로그인
       </Button>
